Formatting code and cascading.

+++ src/main/java/com/spring/stockmanagement/controller/AdminController.java
@@ -1,7 +1,9 @@
 package com.spring.stockmanagement.controller;
 
 
+import com.spring.stockmanagement.Enum.Role;
 import com.spring.stockmanagement.entities.User;
+import com.spring.stockmanagement.helper.Message;
 import com.spring.stockmanagement.service.Interface.CompanyService;
 import com.spring.stockmanagement.service.Interface.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -10,6 +12,7 @@ import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
 
+import javax.servlet.http.HttpSession;
 import java.security.Principal;
 
 @Controller
@@ -23,34 +26,39 @@ public class AdminController {
     private CompanyService companyService;
 
     @ModelAttribute("admin")
-    public User getUser(Principal principal)
-    {
+    public User getUser(Principal principal) {
         String currentUserName = principal.getName();
         User CurrentUser = userService.findByName(currentUserName).get();
         return CurrentUser;
     }
+
     @GetMapping("/")
     public String userDashboard() {
         return "admin/admin_dashboard";
     }
 
     @GetMapping("/users")
-    public String allUsers(Model model)
-    {
-        model.addAttribute("users",userService.getAllUsers());
+    public String allUsers(Model model) {
+        model.addAttribute("users", userService.getAllUsers());
         return "admin/user_list.html";
     }
 
     @GetMapping("/user/delete/{id}")
-    public String deleteUser(@PathVariable int id) {
-        userService.deleteUserById(id);
+    public String deleteUser(@PathVariable int id, HttpSession session) {
+        User user = userService.findById(id);
+        if (user.getRole().equals(Role.STOCKHOLDER)) {
+            session.setAttribute("message", new Message("Unable to delete user with role STOCKHOLDER", "alert-danger"));
+            return "redirect:/admin/users";
+        } else {
+            session.setAttribute("message", new Message("User deleted", "alert-success"));
+            userService.deleteUserById(id);
+        }
         return "redirect:/admin/users";
     }
 
     @GetMapping("/companies")
-    public String listOfCompany(Model model)
-    {
-        model.addAttribute("companies",companyService.getAllCompany());
+    public String listOfCompany(Model model) {
+        model.addAttribute("companies", companyService.getAllCompany());
         return "admin/company_list";
     }
 
@@ -70,16 +78,14 @@ public class AdminController {
     public String updateBook(@PathVariable int id,
                              @ModelAttribute("user") User user,
                              BindingResult bindingResult,
-                             Model model)
-    {
+                             Model model) {
         // save updated user
-        userService.validateUserForUpdate(user,id,bindingResult);
-        if(bindingResult.hasErrors())
-        {
+        userService.validateUserForUpdate(user, id, bindingResult);
+        if (bindingResult.hasErrors()) {
             model.addAttribute(user);
             return "admin/update_admin";
         }
-        userService.updateUser(user,id);
+        userService.updateUser(user, id);
         return "redirect:/admin/";
     }
 }
